% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{assert_dependency}
\alias{assert_dependency}
\title{check if a package is available and return informative message otherwise}
\usage{
assert_dependency(library_name)
}
\description{
check if a package is available and return informative message otherwise
}
\keyword{internal}
