% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_tidiers.R
\name{augment.clmm}
\alias{augment.clmm}
\title{name ordinal_tidiers}
\usage{
\method{augment}{clmm}(x, data = model.frame(x), ...)
}
\description{
the \code{tidy} method for \code{clmm} objects (from the
\code{ordinal} package) lives in the \code{broom} package.
}
