% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.mixed-package.R
\docType{package}
\name{broom.mixed-package}
\alias{broom.mixed}
\alias{broom.mixed-package}
\title{broom.mixed: Tidying Methods for Mixed Models}
\description{
Convert statistical analysis objects from R into tidy data frames,
    so that they can more easily be combined, reshaped and otherwise processed
    with tools like 'dplyr', 'tidyr' and 'ggplot2'. The package provides three
    S3 generics: tidy, which summarizes a model's statistical findings such as
    coefficients of a regression; augment, which adds columns to the original
    data such as predictions, residuals and cluster assignments; and glance, which
    provides a one-row summary of model-level statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/bbolker/broom.mixed}
  \item Report bugs at \url{http://github.com/bbolker/broom.mixed/issues}
}

}
\author{
\strong{Maintainer}: Ben Bolker \email{bolker@mcmaster.ca}

Authors:
\itemize{
  \item David Robinson \email{admiral.david@gmail.com}
}

Other contributors:
\itemize{
  \item Dieter Menne [contributor]
  \item Jonah Gabry [contributor]
  \item Paul Buerkner [contributor]
  \item Christopher Hua [contributor]
  \item Will Petry [contributor]
  \item Josh Wiley [contributor]
  \item Patrick Kennedy [contributor]
}

}
\keyword{internal}
