% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_methods}
\alias{get_methods}
\title{Retrieve all method/class combinations currently provided by the broom.mixed package}
\usage{
get_methods()
}
\description{
Retrieve all method/class combinations currently provided by the broom.mixed package
}
\examples{
print(get_methods(), n = Inf)
}
