% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{show_methods}
\alias{show_methods}
\title{Show all method/class combinations currently provided by the broom.mixed package}
\usage{
show_methods()
}
\description{
Show all method/class combinations currently provided by the broom.mixed package
}
