% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaMM_tidiers.R
\name{spaMM_tidiers}
\alias{spaMM_tidiers}
\alias{tidy.HLfit}
\alias{augment.HLfit}
\alias{glance.HLfit}
\title{Tidying methods for mixed effects models}
\usage{
\method{tidy}{HLfit}(x, effects = c("ran_pars", "fixed"), ...)

\method{augment}{HLfit}(x, ...)

\method{glance}{HLfit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{HLfit}, such as those from \code{fitme},
\code{HLfit}, or \code{corrHLfit}}

\item{effects}{A character vector including one or more of "fixed" (fixed-effect parameters), "ran_pars" (variances and covariances or standard deviations and correlations of random effect terms) or "ran_vals" (conditional modes/BLUPs/latent variable estimates)}

\item{...}{extra arguments (not used)}
}
\value{
All tidying methods return a \code{tibble}.
The structure depends on the method chosen.
}
\description{
These methods tidy the coefficients of mixed effects models, particularly
responses of the \code{HLfit} class
}
\examples{

if (require("spaMM") && require("lme4")) {
  lmm1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
  tidy(lmm1)
  augment(lmm1)
  glance(lmm1)
}
}
