% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarCorr_tidiers.R
\name{tidy.VarCorr.lme}
\alias{tidy.VarCorr.lme}
\title{Tidying VarCorr of Mixed Effect Models}
\usage{
\method{tidy}{VarCorr.lme}(
  object,
  scales = c("sdcor", "vcov"),
  conf.int = FALSE,
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{object}{a VarCorr object}

\item{scales}{see \code{\link{tidy.lme4}}}
}
\description{
Tidying VarCorr of Mixed Effect Models
}
