% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.mixed-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{glance}
\alias{augment}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broom}{\code{\link[broom]{tidy}}, \code{\link[broom]{glance}}, \code{\link[broom]{augment}}}
}}

