% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqmm_tidiers.R
\name{tidy.lqmm}
\alias{tidy.lqmm}
\alias{glance.lqmm}
\title{Tidying methods for lqmm models (EXPERIMENTAL)}
\usage{
\method{tidy}{lqmm}(x, conf.int = FALSE, conf.level = 0.95, ...)

\method{glance}{lqmm}(x, ...)
}
\description{
These methods, suggested by
David Luke Thiessen on \href{https://stackoverflow.com/questions/72514230/is-it-possible-to-use-lqmm-with-a-mira-object}{Stack Exchange}, provide support for linear quantile mixed models. They have not been carefully tested - please
check output carefully and report problems!
}
